third_party/blink/web_tests/platform/generic/fast/dom/shadow/shadow-tree-listener-clearance-expected.txt

PASS firedCount is 1
PASS firedCount is 1
PASS successfullyParsed is true

TEST COMPLETE
